+++ tiledb/sm/cpp_api/config.h
@@ -115,6 +115,13 @@ class TILEDB_EXPORT Config {
    *    The fragment metadata cache size in bytes. Any `uint64_t` value is
    *    acceptable. <br>
    *    **Default**: 10,000,000
+   * - `vfs.max_parallel_ops` <br>
+   *    The maximum number of VFS parallel operations.<br>
+   *    **Default**: number of cores
+   * - `vfs.min_parallel_size` <br>
+   *    The minimum number of bytes in a parallel VFS operation. (Does not
+   *    affect parallel S3 writes.)<br>
+   *    **Default**: 10MB
    * - `vfs.s3.region` <br>
    *    The S3 region, if S3 is enabled. <br>
    *    **Default**: us-east-1
@@ -128,9 +135,11 @@ class TILEDB_EXPORT Config {
    *    The S3 use of virtual addressing (`true` or `false`), if S3 is
    *    enabled. <br>
    *    **Default**: true
-   * - `vfs.s3.file_buffer_size` <br>
-   *    The file buffer size (in bytes) used in S3 writes, if S3 is enables. Any
-   *    `uint64_t` value is acceptable. <br>
+   * - `vfs.s3.multipart_part_size` <br>
+   *    The part size (in bytes) used in S3 multipart writes, if S3 is enabled.
+   *    Any `uint64_t` value is acceptable. Note: `vfs.s3.multipart_part_size *
+   *    vfs.max_parallel_ops` bytes will be buffered before issuing multipart
+   *    uploads in parallel. <br>
    *    **Default**: 5*1024*1024
    * - `vfs.s3.connect_timeout_ms` <br>
    *    The connection timeout in ms. Any `long` value is acceptable. <br>