+++ tiledb/sm/storage_manager/config.cc
@@ -170,8 +170,8 @@ Status Config::set(const std::string& param, const std::string& value) {
     RETURN_NOT_OK(set_vfs_s3_endpoint_override(value));
   } else if (param == "vfs.s3.use_virtual_addressing") {
     RETURN_NOT_OK(set_vfs_s3_use_virtual_addressing(value));
-  } else if (param == "vfs.s3.file_buffer_size") {
-    RETURN_NOT_OK(set_vfs_s3_file_buffer_size(value));
+  } else if (param == "vfs.s3.multipart_part_size") {
+    RETURN_NOT_OK(set_vfs_s3_multipart_part_size(value));
   } else if (param == "vfs.s3.connect_timeout_ms") {
     RETURN_NOT_OK(set_vfs_s3_connect_timeout_ms(value));
   } else if (param == "vfs.s3.connect_max_tries") {
@@ -266,10 +266,11 @@ Status Config::unset(const std::string& param) {
                                                                "false");
     param_values_["vfs.s3.use_virtual_addressing"] = value.str();
     value.str(std::string());
-  } else if (param == "vfs.s3.file_buffer_size") {
-    vfs_params_.s3_params_.file_buffer_size_ = constants::s3_file_buffer_size;
-    value << vfs_params_.s3_params_.file_buffer_size_;
-    param_values_["vfs.s3.file_buffer_size"] = value.str();
+  } else if (param == "vfs.s3.multipart_part_size") {
+    vfs_params_.s3_params_.multipart_part_size_ =
+        constants::s3_multipart_part_size;
+    value << vfs_params_.s3_params_.multipart_part_size_;
+    param_values_["vfs.s3.multipart_part_size"] = value.str();
     value.str(std::string());
   } else if (param == "vfs.s3.connect_timeout_ms") {
     vfs_params_.s3_params_.connect_timeout_ms_ =
@@ -359,8 +360,8 @@ void Config::set_default_param_values() {
   param_values_["vfs.s3.use_virtual_addressing"] = value.str();
   value.str(std::string());
 
-  value << vfs_params_.s3_params_.file_buffer_size_;
-  param_values_["vfs.s3.file_buffer_size"] = value.str();
+  value << vfs_params_.s3_params_.multipart_part_size_;
+  param_values_["vfs.s3.multipart_part_size"] = value.str();
   value.str(std::string());
 
   value << vfs_params_.s3_params_.connect_timeout_ms_;
@@ -459,10 +460,10 @@ Status Config::set_vfs_s3_use_virtual_addressing(const std::string& value) {
   return Status::Ok();
 }
 
-Status Config::set_vfs_s3_file_buffer_size(const std::string& value) {
+Status Config::set_vfs_s3_multipart_part_size(const std::string& value) {
   uint64_t v;
   RETURN_NOT_OK(utils::parse::convert(value, &v));
-  vfs_params_.s3_params_.file_buffer_size_ = v;
+  vfs_params_.s3_params_.multipart_part_size_ = v;
 
   return Status::Ok();
 }